Added single page for both Create action and Edit action in Positions,Employees,Articles folders
