+++ Pages/ArticlesF/Create.cshtml
@@ -1,7 +1,7 @@
 ﻿@page
 @model MYChamp.Pages.ArticlesF.CreateModel
 @{
-    ViewData["Title"] = "Create page";
+    ViewData["Title"] = Model.Article.Id==0 ? "Create page" : "Edit Page" ;
 }
 
 <style>
@@ -14,7 +14,10 @@
 </style>
 
 <form method="post" enctype="multipart/form-data">
+    <input asp-for="Article.Id" hidden />
     <div class="container-fluid p-3">
+        <div class="row">
+            <div class="@(Model.Article.Id == 0 ? "col-12" : "col-9")">
         <div class="form-floating mb-3 mt-3">
             <input asp-for="Article.Title" class="form-control" placeholder="Title" />
             <label asp-for="Article.Title">Title</label>
@@ -71,9 +74,23 @@
 
         <div class="row mb-1">
             <div class="col-6 col-md-3">
-                <button type="submit" class="btn btn-primary form-control">Create</button>
+                <button type="submit" class="btn btn-primary form-control">@((Model.Article.Id == 0) ? "Add" : "Save")</button>
             </div>
-        </div>
+        </div>    
+     </div>   
+        @if(Model.Article.Id!=0)
+        {
+            <div class="col-3">
+                @if (Model.Article.CoverImageData != null && Model.Article.CoverImageData.Length > 0)
+                {
+                    <img src="data:image/jpeg;base64,@Convert.ToBase64String(Model.Article.CoverImageData)" width="100%" style="border-radius: 5px; border: 1px solid #bbb9b9" />
+                }
+            </div>
+        }
+    </div>  
+    </div>
+    <div>
+        <input type="hidden" asp-for="Article.UserEmail" />
     </div>
 </form>
