--- Pages/ArticlesF/Edit.cshtml
@@ -1,106 +1,0 @@
-﻿@page
-@model MYChamp.Pages.ArticlesF.EditModel
-@{
-    ViewData["Title"] = "Edit page";
-}
-
-<form method="post" enctype="multipart/form-data">
-    <div class="row">
-        <div class="col-9">
-            <input asp-for="Article.Id" hidden />
-
-            <div class="form-floating mb-3">
-                <input asp-for="Article.Title" class="form-control" id="title" />
-                <label for="title">Title</label>
-                <span asp-validation-for="Article.Title" class="text-danger"></span>
-            </div>
-
-            <div class="form-floating mb-3">
-                <textarea asp-for="Article.Description" class="form-control" id="description" style="height: 100px;"></textarea>
-                <label for="description">Description</label>
-                <span asp-validation-for="Article.Description" class="text-danger"></span>
-            </div>
-
-            <div class="form-floating mb-3">
-                <textarea asp-for="Article.Relevancy" class="form-control" id="relevancy" style="height: 150px;"></textarea>
-                <label for="relevancy">Relevancy</label>
-                <span asp-validation-for="Article.Relevancy" class="text-danger"></span>
-            </div>
-
-            <div class="form-floating mb-3">
-                <input type="url" asp-for="Article.MediaLinks" class="form-control" id="mediaLinks" />
-                <label for="mediaLinks">Media Links</label>
-                <span asp-validation-for="Article.MediaLinks" class="text-danger"></span>
-            </div>
-
-            <div class="form-floating mb-3">
-                <input type="file" asp-for="Article.CoverPath" class="form-control" id="coverImage" />
-                <label for="coverImage">Cover Image</label>
-            </div>
-
-            <div class="form-floating mb-3">
-                <input asp-for="Article.UserEmail" class="form-control" id="userEmail" readonly />
-                <label for="userEmail">User Email</label>
-                <span asp-validation-for="Article.UserEmail" class="text-danger"></span>
-            </div>
-
-            <div class="form-floating mb-3">
-                <input asp-for="Article.Author" class="form-control" id="author" />
-                <label for="author">Author</label>
-                <span asp-validation-for="Article.Author" class="text-danger"></span>
-            </div>
-
-            <div class="form-floating mb-3">
-                <input asp-for="Article.AuthorPosition" class="form-control" id="authorPosition" />
-                <label for="authorPosition">Author Position</label>
-                <span asp-validation-for="Article.AuthorPosition" class="text-danger"></span>
-            </div>
-
-            <div class="form-floating mb-3">
-                <input asp-for="Article.Category" class="form-control" id="category" />
-                <label for="category">Category</label>
-                <span asp-validation-for="Article.Category" class="text-danger"></span>
-            </div>
-
-            <div class="row mb-1">
-                <div class="col-6 col-md-3">
-                    <button type="submit" class="btn btn-primary form-control">Update</button>
-                </div>
-            </div>
-        </div>
-
-        <div class="col-3">
-            @if (Model.Article.CoverImageData != null && Model.Article.CoverImageData.Length > 0)
-            {
-                <img src="data:image/jpeg;base64,@Convert.ToBase64String(Model.Article.CoverImageData)" width="100%" style="border-radius: 5px; border: 1px solid #bbb9b9" />
-            }
-        </div>
-    </div>
-</form>
-
-@section Scripts {
-    <script>
-        $(document).ready(function () {
-            $('form').submit(function (event) {
-                event.preventDefault();
-
-                var form = $(this)[0];
-                var formData = new FormData(form);
-
-                $.ajax({
-                    type: 'POST',
-                    url: $(this).attr('action'),
-                    data: formData,
-                    contentType: false,
-                    processData: false,
-                    success: function (result) {
-                        window.parent.location.reload();
-                    },
-                    error: function (xhr, status, error) {
-                        alert('An error occurred: ' + error);
-                    }
-                });
-            });
-        });
-    </script>
-}
