+++ Pages/ArticlesF/Index.cshtml
@@ -67,7 +67,7 @@
             <tr>
                 <th>Title</th>
                 <th>Category</th>
-                <th></th>
+                <th>Actions</th>
             </tr>
         </thead>
         <tbody>
@@ -82,7 +82,7 @@
                                @*  <a asp-page="/ArticlesF/Edit" asp-route-id="@item.Id" class="btn btn-primary">
                                     <i class="bi bi-pencil"></i> Edit
                                 </a> *@
-                                <a href="javascript:void(0);" class="btn btn-primary" onclick="openCommoniFrameModal('/ArticlesF/Edit?id=@item.Id', 'Edit Article')">
+                                <a href="javascript:void(0);" class="btn btn-primary" onclick="openCommoniFrameModal('/ArticlesF/Create?id=@item.Id', 'Edit Article')">
                                     <i class="bi bi-pencil"></i> Edit
                                 </a>
