+++ Pages/ArticlesF/Index.cshtml.cs
@@ -1,3 +1,4 @@
+
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using MYChamp.DbContexts;
