+++ Pages/Employees/Create.cshtml
@@ -1,9 +1,12 @@
-﻿@page
+﻿@page "{id?}"
 @model MYChamp.Pages.Employees.CreateModel
-
-<h2 class="pt-5 text-center">New Employee</h2>
+{
+    ViewData["Title"]=Model.Employee.id==0 ? "Add New Employee" : "Edit Employee Details";
+}
+<h2 class="pt-5 text-center">@ViewData["Title"]</h2>
 
 <form method="post" class="container-fluid col-10">
+     <input type="hidden" asp-for="Employee.Id" />
     <div class="form-floating mb-3">
         <input asp-for="Employee.EmployeeId" class="form-control" id="EmployeeId" placeholder="Employee ID" />
         <label for="EmployeeId">Employee ID </label>
@@ -33,7 +36,7 @@
         <label for="Reporting">Reporting Manager</label>
         <span asp-validation-for="Employee.ReportingManagerId" class="text-danger"></span>
     </div>
-    <button type="submit" class="btn btn-primary mt-1">Create</button>
+    <button type="submit" class="btn btn-primary mt-1">@((Model.Employee.Id == 0) ? "Add" : "Save")</button>
     <a asp-page="./Index" class="btn btn-secondary mt-1">Cancel</a>
 </form>
