+++ Pages/Positions/Create.cshtml.cs
@@ -1,5 +1,6 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using Microsoft.EntityFrameworkCore;
 using MYChamp.DbContexts;
 using MYChamp.Models;
 
@@ -17,8 +18,24 @@ namespace MYChamp.Pages.Positions
         [BindProperty]
         public Position Position { get; set; }
 
-        public IActionResult OnGet()
+        public async Task<IActionResult> OnGetAsync(int? id)
         {
+            if (id == null)
+            {
+               
+                Position = new Position();
+            }
+            else
+            {
+               
+                Position = await _context.Positions.FindAsync(id);
+
+                if (Position == null)
+                {
+                    return NotFound();
+                }
+            }
+
             return Page();
         }
 
@@ -29,7 +46,45 @@ namespace MYChamp.Pages.Positions
                 return Page();
             }
 
-            _context.Positions.Add(Position);
+            if (Position.Id == 0)
+            {
+                _context.Positions.Add(Position);
+            }
+            else
+            {
+               
+                var positionToUpdate = await _context.Positions
+                                                      .AsNoTracking()
+                                                      .FirstOrDefaultAsync(p => p.Id == Position.Id);
+                if (positionToUpdate == null)
+                {
+                    return NotFound();
+                }
+
+                int differenceInLeaves = Position.AllottedLeaves - positionToUpdate.AllottedLeaves;
+
+                if (differenceInLeaves != 0)
+                {
+                    var employees = _context.Employees
+                                            .Where(e => e.PositionId == Position.Id && !e.IsActive)
+                                            .ToList();
+
+                    foreach (var employee in employees)
+                    {
+                        employee.RemainingLeaves += differenceInLeaves;
+
+                        if (employee.RemainingLeaves < 0)
+                        {
+                            employee.RemainingLeaves = 0;
+                        }
+
+                        _context.Entry(employee).State = EntityState.Modified;
+                    }
+                }
+
+                _context.Positions.Update(Position);
+            }
+
             await _context.SaveChangesAsync();
 
             return RedirectToPage("./Index");
