--- Pages/Positions/Edit.cshtml
@@ -1,23 +1,0 @@
-﻿@page
-@model MYChamp.Pages.Positions.EditModel
-@{
-    ViewData["Title"] = "Edit Position";
-}
-
-<h2 class="pt-5 text-center">Edit Position</h2>
-
-<form method="post" class="container-fluid col-lg-10">
-    <input type="hidden" asp-for="Position.Id" />
-    <div class="form-floating mb-3">
-        <input asp-for="Position.Name" class="form-control" id="PositionName" placeholder="Position Name" />
-        <label for="PositionName"> Position Name</label>
-        <span asp-validation-for="Position.Name" class="text-danger"></span>
-    </div>
-    <div class="form-floating mb-3">
-        <input asp-for="Position.AllottedLeaves" class="form-control"  id="PositionLeaves" placeholder="Alloted Leaves" />
-        <label for="PositionLeaves">Alloted Leaves </label>
-        <span asp-validation-for="Position.AllottedLeaves" class="text-danger"></span>
-    </div>
-    <button type="submit" class="btn btn-primary mt-2">Save</button>
-    <a asp-page="./Index" class="btn btn-secondary mt-2">Cancel</a>
-</form>
